+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/controllers/SchoolController.java
@@ -1,6 +1,9 @@
 package com.nimio2024.nimio2024_tfg_polmorenocastelar.controllers;
 
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.dto.SchoolDTO;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.exceptions.CenterDoNotExistException;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.exceptions.SchoolAlreadyExistsException;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.exceptions.SchoolDoNotExistException;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Center;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Course;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.School;
@@ -22,35 +25,41 @@ public class SchoolController {
         this.centerService = centerService;
     }
 
-    public School createSchool(SchoolDTO schoolDTO, Long centerId) {
+    public School createSchool(SchoolDTO schoolDTO, Long centerId) throws SchoolAlreadyExistsException, CenterDoNotExistException {
         School school = new School(schoolDTO);
         if(schoolService.getSchoolByName(school.getSchoolName()).isPresent()) {
-            //throw new IllegalArgumentException("School already exists"); //TODO: THROW O SYSERR?
-            System.err.println("School already exists, check the name");
-            return null;
+            throw new SchoolAlreadyExistsException("School with name [" + schoolDTO.getSchoolName() + "] already exists");
         }
         Center center = centerService.getCenterById(centerId);
         if(center == null){
-            //TODO: THROW O SYSERR?
-            System.err.println("Center not found, check the center id");
-            return null;
+            throw new CenterDoNotExistException("Center with id " + centerId + " not found, check the center id");
         }
         school.setCenter(center);
         centerService.addSchool(center, school);
         return schoolService.saveSchool(school);
     }
 
-    public School getSchool(Long schoolId){
-        return schoolService.getSchoolById(schoolId);
+    public School getSchool(Long schoolId) throws SchoolDoNotExistException {
+        School school = schoolService.getSchoolById(schoolId);
+        if(school == null){
+            throw new SchoolDoNotExistException("School with id " + schoolId + " not found, check the school id");
+        }
+        return school;
     }
 
-    public List<Course> getSchoolCourses(Long schoolId) {
+    public List<Course> getSchoolCourses(Long schoolId) throws SchoolDoNotExistException {
         School school = schoolService.getSchoolById(schoolId);
+        if(school == null){
+            throw new SchoolDoNotExistException("School with id " + schoolId + " not found, check the school id");
+        }
         return school.getCourseList();
     }
 
-    public School updateSchool(Long schoolId, SchoolDTO schoolDTO) {
+    public School updateSchool(Long schoolId, SchoolDTO schoolDTO) throws SchoolDoNotExistException {
         School school = schoolService.getSchoolById(schoolId);
+        if(school == null){
+            throw new SchoolDoNotExistException("School with id " + schoolId + " not found, check the school id");
+        }
         school.setSchoolName(schoolDTO.getSchoolName());
         school.setSchoolLocationLatitude(schoolDTO.getSchool_location_latitude());
         school.setSchoolLocationLongitude(schoolDTO.getSchool_location_longitude());
@@ -61,19 +70,28 @@ public class SchoolController {
         schoolService.deleteSchool(schoolId);
     }
 
-    public String getCenterNameBySchoolId(Long schoolId) {
+    public String getCenterNameBySchoolId(Long schoolId) throws SchoolDoNotExistException {
         School school = schoolService.getSchoolById(schoolId);
+        if(school == null){
+            throw new SchoolDoNotExistException("School with id " + schoolId + " not found, check the school id");
+        }
         return school.getCenter().getCenterName();
     }
 
-    public School updateSchoolName(Long schoolId, String schoolName) {
+    public School updateSchoolName(Long schoolId, String schoolName) throws SchoolDoNotExistException {
         School school = schoolService.getSchoolById(schoolId);
+        if(school == null){
+            throw new SchoolDoNotExistException("School with id " + schoolId + " not found, check the school id");
+        }
         school.setSchoolName(schoolName);
         return schoolService.saveSchool(school);
     }
 
-    public List<Teacher> getTeachersFromSchool(Long schoolId) {
+    public List<Teacher> getTeachersFromSchool(Long schoolId) throws SchoolDoNotExistException {
         School school = schoolService.getSchoolById(schoolId);
+        if(school == null){
+            throw new SchoolDoNotExistException("School with id " + schoolId + " not found, check the school id");
+        }
         return school.getTeachers();
     }
 }